Update from repo

+++ core-cube/src/main/java/org/apache/kylin/cube/cuboid/algorithm/CuboidStatsUtil.java
@@ -29,8 +29,8 @@ import java.util.TreeSet;
 
 import org.apache.kylin.common.util.Pair;
 
-import com.google.common.collect.Lists;
-import com.google.common.collect.Maps;
+import org.apache.kylin.shaded.com.google.common.collect.Lists;
+import org.apache.kylin.shaded.com.google.common.collect.Maps;
 
 public class CuboidStatsUtil {
 
@@ -125,7 +125,7 @@ public class CuboidStatsUtil {
                     nEffective++;
                 }
             }
-            
+
             if (nEffective != 0)
                 srcCuboidsStats.put(cuboid, totalEstRowCount / nEffective);
             else
@@ -349,7 +349,7 @@ public class CuboidStatsUtil {
     }
 
     private static double calculateRollupRatio(Pair<Long, Long> rollupStats) {
-        double rollupInputCount = rollupStats.getFirst() + rollupStats.getSecond();
+        double rollupInputCount = (double) rollupStats.getFirst() + rollupStats.getSecond();
         return rollupInputCount == 0 ? 0 : 1.0 * rollupStats.getFirst() / rollupInputCount;
     }
 }
